external/wakaama : change type to fix a error.
This commit is to fix a logical error.
Expression 'len < ((1 << 16) + 269)' is always true,
which may be caused by a logical error.

+++ external/wakaama/core/er-coap-13/er-coap-13.c
@@ -388,7 +388,7 @@ size_t coap_serialize_get_size_tcp(void *packet)
 {
 	coap_packet_t *coap_pkt = (coap_packet_t *) packet;
 
-	uint16_t len = 0;
+	uint32_t len = 0;
 	uint16_t options_len = 0;
 
 	size_t length = 0;
